requirements.txt
robotframework==4.0.3
python-dotenv==0.15.0
requests==2.25.0